refactor(types): restore centralized type system architecture
Fixes problematic revert from commit c9952c0 that undid the type centralization
work from commit 945fb47. This commit fully restores the clean separation of
concerns between types and implementation code in src/lib/core/ module.

- **REMOVED** `/src/lib/core/types.ts` (403 lines of duplicate types)
- **ENHANCED** `/src/lib/types/providers.ts` with missing O1 models, `APIVersions` enum, and `DEFAULT_MODEL_ALIASES`
- **UPDATED** 12 files to use centralized imports: `../core/types.js` → `../types/providers.js|generateTypes.js`

- Types: Centralized in `/src/lib/types/` (definitions only)
- Core: Implementation in `/src/lib/core/` (logic only)
- Zero type definitions remain in core modules
