+++ src/lib/types/providers.ts
@@ -59,6 +59,8 @@ export enum OpenAIModels {
   GPT_4O = "gpt-4o",
   GPT_4O_MINI = "gpt-4o-mini",
   GPT_3_5_TURBO = "gpt-3.5-turbo",
+  O1_PREVIEW = "o1-preview",
+  O1_MINI = "o1-mini",
 }
 
 /**
@@ -121,6 +123,29 @@ export enum AnthropicModels {
   CLAUDE_3_SONNET = "claude-3-sonnet-20240229",
   CLAUDE_3_OPUS = "claude-3-opus-20240229",
   CLAUDE_3_HAIKU = "claude-3-haiku-20240307",
+}
+
+/**
+ * API Versions for various providers
+ */
+export enum APIVersions {
+  // Azure OpenAI API versions
+  AZURE_LATEST = "2025-04-01-preview",
+  AZURE_STABLE = "2024-10-21",
+  AZURE_LEGACY = "2023-12-01-preview",
+
+  // OpenAI API versions
+  OPENAI_CURRENT = "v1",
+  OPENAI_BETA = "v1-beta",
+
+  // Google AI API versions
+  GOOGLE_AI_CURRENT = "v1",
+  GOOGLE_AI_BETA = "v1beta",
+
+  // Anthropic API versions
+  ANTHROPIC_CURRENT = "2023-06-01",
+
+  // Other provider versions can be added here
 }
 
 /**
@@ -532,6 +557,30 @@ export type AISDKGenerateResult = GenerateResult & {
   }>;
   [key: string]: unknown;
 };
+
+/**
+ * Default model aliases for easy reference
+ */
+export const DEFAULT_MODEL_ALIASES = {
+  // Latest recommended models per provider
+  LATEST_OPENAI: OpenAIModels.GPT_4O,
+  FASTEST_OPENAI: OpenAIModels.GPT_4O_MINI,
+  LATEST_ANTHROPIC: AnthropicModels.CLAUDE_3_5_SONNET,
+  FASTEST_ANTHROPIC: AnthropicModels.CLAUDE_3_5_HAIKU,
+  LATEST_GOOGLE: GoogleAIModels.GEMINI_2_5_PRO,
+  FASTEST_GOOGLE: GoogleAIModels.GEMINI_2_5_FLASH,
+
+  // Best models by use case
+  BEST_CODING: AnthropicModels.CLAUDE_3_5_SONNET,
+  BEST_ANALYSIS: GoogleAIModels.GEMINI_2_5_PRO,
+  BEST_CREATIVE: AnthropicModels.CLAUDE_3_5_SONNET,
+  BEST_VALUE: GoogleAIModels.GEMINI_2_5_FLASH,
+} as const;
+
+/**
+ * @deprecated Use DEFAULT_MODEL_ALIASES instead. Will be removed in future version.
+ */
+export const ModelAliases = DEFAULT_MODEL_ALIASES;
 
 /**
  * Default provider configurations
