+++ src/lib/types/streamTypes.ts
@@ -2,7 +2,7 @@ import type { Tool } from "ai";
 import type { ValidationSchema, StandardRecord } from "./typeAliases.js";
 import type { ProviderConfig } from "./providers.js";
 import type { TextContent, ImageContent } from "./content.js";
-import type { AIProviderName, AnalyticsData } from "../core/types.js";
+import type { AIProviderName, AnalyticsData } from "../types/index.js";
 import type { TokenUsage } from "./analytics.js";
 import type { EvaluationData } from "../index.js";
 import type {
